requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit

diffusers[training]
einops>=0.8.0
blosc>=1.11.2
